refactor: extract task controls from AgentWorkspace
